+++ l10n/tr.js
@@ -24,6 +24,7 @@ OC.L10N.register(
         'Track' : 'Parça',
         'File' : 'Dosya',
         'Title' : 'Başlık',
+        'Titles': 'Başlıklar',
         'Subtitle' : 'Altbaşlık',
         'Artist' : 'Sanatçı',
         'Album Artist' : 'Albüm Sanatçısı',